api/requirements.txt
fastapi>=0.109.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0